refactor: move audit test-only options to export_test.go
WithClock and WithRedactor are test seams whose only consumer is the
sibling audit_test package; per the repo convention they live in
export_test.go, not the gated production file.
